Create new TheiaContextMenuService per widget
Otherwise the last opened GLSP widget connects to the same
TheiaContextMenuService over and over again so that all context menu
actions are performed in the action dispatcher of the last widget.

To make this more convenient for clients, we provide a factory and a
function for hooking up the context menu service.

https://github.com/eclipse-glsp/glsp/issues/57