#377 Update prettier config
Update to latest @eclipse-glsp/config version. This version ships with prettier as dependency which enables to configure a workspace-local prettier version for the vscode prettier extension. This ensures that all developers are using the same prettier version.

.settings:
- Remove prettier as general default formatter to avoid unintended usage of prettier - for languages like java.
- Configure prettier as default formatter for CSS
- Set the prettier path to "node_modules/prettier"

Reformat all code:
Reformat code with "yarn prettier --config .prettierrc ./ --write" in root to ensure that all files for which prettier is set as default formatter are properly formatted.