Fix duplicate viewport change on navigation
If we directly open a digram and navigate to an element
we now get two viewport changes (one restores the viewport
and one navigates to the navigation target.
This can be reproduced by closing the editor on a particular
viewport and then using the Theia command
"Open example1.wf task "Push"".

To avoid that, we let the diagram manager decide whether the
widget should restore the viewport. Imho this gives more
flexibility anyway.

https://github.com/eclipse-glsp/glsp/issues/475

Change-Id: I6f6d5f503a63682beb4e770f25f437c133226d78