GLSP-1308: Ensure that hidden div is properly removed on close
Ensure that the hidden div is properly removed when a diagram is closed.
The hidden div is purposely part of the document body and not the diagram root itself
This ensures that the hidden computation also works as expected if the diagram is currently hidden 
e.g. because another tab is in focus
However, this also means that we have to cleanup the hidden div when the diagram closes i.e. is detached.
Contributed on behalf of STMicroelectronics
Fixes https://github.com/eclipse-glsp/glsp/issues/1308